refactor: review code LayananDk API

+++ Bilreg.Domain/AdmisiContext/LayananSub/LayananDkAgg/LayananDkModel.cs
@@ -1,27 +1,25 @@
 ﻿namespace Bilreg.Domain.AdmisiContext.LayananSub.LayananDkAgg
 {
     public class LayananDkModel(
-     string  layananDkId,
-     string  layananDkName,
-     decimal rawatInapCode,
-     decimal rawatJalanCode,
-     decimal kesehatanJiwaCode,
-     decimal bedahCode,
-     decimal rujukanCode,
-     decimal kunjunganRumahCode,
-     decimal layananSubCode
-     
-        ) :ILayananDkKey
+        string  layananDkId,
+        string  layananDkName,
+        int rawatInapCode,
+        int rawatJalanCode,
+        int kesehatanJiwaCode,
+        int bedahCode,
+        int rujukanCode,
+        int kunjunganRumahCode,
+        int layananSubCode) :ILayananDkKey
     {
         public string LayananDkId { get; protected set; } = layananDkId ;
         public string LayananDkName { get; protected set; } = layananDkName ;
-        public decimal RawatInapCode { get; protected set; } = rawatInapCode ;
-        public decimal RawatJalanCode { get; protected set; } = rawatJalanCode ;
-        public decimal KesehatanJiwaCode { get; protected set; } = kesehatanJiwaCode ;
-        public decimal BedahCode { get; protected set; } = bedahCode ;
-        public decimal RujukanCode { get; protected set; } = rujukanCode ;
-        public decimal KunjunganRumahCode { get; protected set; } = kunjunganRumahCode ;
-        public decimal LayananSubCode { get; protected set; } = layananSubCode;
+        public int RawatInapCode { get; protected set; } = rawatInapCode ;
+        public int RawatJalanCode { get; protected set; } = rawatJalanCode ;
+        public int KesehatanJiwaCode { get; protected set; } = kesehatanJiwaCode ;
+        public int BedahCode { get; protected set; } = bedahCode ;
+        public int RujukanCode { get; protected set; } = rujukanCode ;
+        public int KunjunganRumahCode { get; protected set; } = kunjunganRumahCode ;
+        public int LayananSubCode { get; protected set; } = layananSubCode;
     }
 
     public interface ILayananDkKey
